Use separate classes for cntainer migrations
